Move navigation components, fix e2e
